Change users.credits from INTEGER to NUMERIC(10,2) for decimal values
Database Migration:
- Created migration to alter users.credits column type
- Changed from INTEGER to NUMERIC(10,2) to support decimal credits
- Added check constraint to prevent negative credits
- Updated default value to 0.00

Backend Code Changes:
- Updated User model: credits now uses DataTypes.DECIMAL(10,2)
- Removed Math.floor conversion in credit-service.js
- Credits are now stored with decimal precision (e.g., 0.1, 0.5, 99.75)

Cleanup:
- Removed excessive console.log statements from translations.js entity-stats route

This allows fractional credit charges for services like:
- Standard translations: 0.1 credits
- CMS blocks: 0.2 credits
- CMS pages: 0.5 credits
- Custom domain daily: 0.5 credits

Run the migration SQL on Supabase to apply the schema change.

+++ backend/src/models/User.js
@@ -68,9 +68,9 @@ const User = sequelize.define('User', {
     defaultValue: 'agency'
   },
   credits: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.DECIMAL(10, 2),
     allowNull: true,
-    defaultValue: 0
+    defaultValue: 0.00
   }
 }, {
   tableName: 'users',
